Renamed Program.cs to Apphost.cs
